gamepad array connectivity independance achieved
If you unplug one gamepad, none of the other gampads input will change (they remain in the same place in the array).

If you plug the gamepad again, it will remain in its position of the array, thus if player 1 unplugs the gamepad, and plugs it in again, that gamepad will still be player 1.