Ui timer finished + storm movement based in tiles
The UI Timer should work perfectly both to have a countdown of how much time it takes until the storm starts to move and how much time the storm will be moving (calculated for each phase automatically!).

Also, the zone now moves in perfect amount of tiles as described in the variables and in the xml, meaning now in each phase you input how much time the storm waits until it moves, and how much tiles it will advance until starting the next phase.